requirements.txt
dbt-postgres>=1.5.0
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
clickhouse-connect>=0.6.0
python-dotenv>=1.0.0
